+++ pymc/distributions/__init__.py
@@ -21,6 +21,7 @@ from pymc.distributions.logprob import (  # isort:skip
 )
 
 from pymc.distributions.bound import Bound
+from pymc.distributions.censored import Censored
 from pymc.distributions.continuous import (
     AsymmetricLaplace,
     Beta,
@@ -187,6 +188,7 @@ __all__ = [
     "Rice",
     "Moyal",
     "Simulator",
+    "Censored",
     "CAR",
     "PolyaGamma",
     "logpt",